Changed get notebook options from nativeEditor.ts to nativeEditorProvider.ts to make it callable from baseIntellisenseProvider.ts, this might be undone later as the issue is not yet resolved.
