+++ src/test/datascience/intellisense.unit.test.ts
@@ -18,7 +18,7 @@ import {
     IInteractiveWindowMapping,
     InteractiveWindowMessages
 } from '../../client/datascience/interactive-common/interactiveWindowTypes';
-import { IInteractiveWindowListener, IInteractiveWindowProvider, IJupyterExecution } from '../../client/datascience/types';
+import { IInteractiveWindowListener, IInteractiveWindowProvider, IJupyterExecution, INotebookEditorProvider } from '../../client/datascience/types';
 import { MockAutoSelectionService } from '../mocks/autoSelector';
 import { MockLanguageClient } from './mockLanguageClient';
 
@@ -34,6 +34,7 @@ suite('DataScience Intellisense Unit Tests', () => {
     let fileSystem: TypeMoq.IMock<IFileSystem>;
     let jupyterExecution: TypeMoq.IMock<IJupyterExecution>;
     let interactiveWindowProvider: TypeMoq.IMock<IInteractiveWindowProvider>;
+    let nativeEditorProvider: TypeMoq.IMock<INotebookEditorProvider>;
     const pythonSettings = new class extends PythonSettings {
         public fireChangeEvent() {
             this.changed.fire();
@@ -51,6 +52,7 @@ suite('DataScience Intellisense Unit Tests', () => {
         fileSystem = TypeMoq.Mock.ofType<IFileSystem>();
         jupyterExecution = TypeMoq.Mock.ofType<IJupyterExecution>();
         interactiveWindowProvider = TypeMoq.Mock.ofType<IInteractiveWindowProvider>();
+        nativeEditorProvider = TypeMoq.Mock.ofType<INotebookEditorProvider>();
 
         pythonSettings.jediEnabled = false;
         languageServer.setup(l => l.start(TypeMoq.It.isAny(), TypeMoq.It.isAny())).returns(() => Promise.resolve());
@@ -66,7 +68,9 @@ suite('DataScience Intellisense Unit Tests', () => {
             configService.object,
             fileSystem.object,
             jupyterExecution.object,
-            interactiveWindowProvider.object);
+            interactiveWindowProvider.object,
+            nativeEditorProvider.object
+        );
     });
 
     function sendMessage<M extends IInteractiveWindowMapping, T extends keyof M>(type: T, payload?: M[T]): Promise<void> {
